Correction of sphere collision on brick game
Start becomes awake in order to cache rigidbody component. If Start
remains, the rigidbody is accessed from somewhere while not yet assigned
resulting in a null reference.
Transform and rigidbody are cached for performance.
The sphere does not have a fake gravity but checks for the velocity and
get the angle with Vector3.right. If the angle is too small the ball is
diverted to avoid horizontal movements.
On collision, the sphere gets a new velocity based on collision
resolution formula.
SphereSounds gets a reference in the main sphere script to play sounds
on collision.
SphereSounds has a new method instead of the collision to play sounds.
The method is called when collision is detected from the main script
avoiding double call of the callback method.